Create README.md
The objective of the experiment is to design a DC-DC converter circuit that will have to be able to change the duty cycle on its own, depending on the input voltage level and will provide a 
fixed DC output for an AC input. For an AC supply of 100V Peak at 60Hz as input we need to have a DC output of 50V (+/- 3.5V) at the steady state region. To get the fixed DC output from an AC input, we have to design a converter and according to the requirements, it should be a DC-DC converter. We need to find out the value of components to design a circuit where the components value has to be at least 10 times the minimum requirement and depending on the input voltage level the circuit should have the ability to change the duty cycle on its own to deliver a fixed DC output.